refactor: next to OccTable we also need the concept of SymbolVectors

+++ src/fmindex-collection/occtable/Bitvector.h
@@ -163,7 +163,7 @@ struct OccTable {
         return a;
     }
 
-    uint64_t symbol(uint64_t idx) const {
+    uint8_t symbol(uint64_t idx) const {
         for (uint64_t i{0}; i < Sigma-1; ++i) {
             if (bitvector[i].value(idx)) {
                 return i;
